[Experiment] Speed up cli.
We are using nedb as the backend storage, which is designed to
bump all data into memory when starting thus further CRUD will
be faster.

However our cli doesn't taste this profit if we are running in
cli way. For each time running we have to irepeatly bump the db
into memory, which is definitely a waste since we are going to
exit sooner anyway.

This experiemnt is using a server-client mode to fix this: the
server is a standalone process running as a daemon, which still
bump the data into memory as usual but live forever. The client
keeps the same as original cli while will get data from server
instead.

Signed-off-by: Eric Wang <lei.wang2@emc.com>